requirements.txt
torch==2.0.1
torchvision==0.15.2
pygame==2.5.2
matplotlib==3.7.1
Pillow==9.5.0
numpy==1.24.3
